# Implement ~20x faster SkyCoord getitem using cached template