This addresses issue #3323 by implementing `SkyCoord.__getitem__` using a cached scalar template SkyCoord and then filling in the representation component data and attributes.

On my mac this speeds up getitem and iteration by a factor of about 20.