@taldcroft 2017-11-25
Ugh, there's a problem when changing representations....