@taldcroft 2017-11-26
@mhvk @astrofrog - passing tests locally now.