@taldcroft 2017-12-01
> So just to be clear, even with the deepcopy you are getting a 20x speedup? (if so - wow!). Also I guess the deepcopy does return a different SkyCoord for each item?

Yes and yes.