@taldcroft 2017-12-01
Though I see now the uniqueness of returned values is not tested.  Will do that, but in development I extensively tested by doing `list(sky_coords)` and seeing that they were all different.

BTW, without the uniqueness requirement one can have a `fastiter()` method that is something like a few hundred times faster since you avoid the deepcopy.  @mhvk was opposed to "polluting" the API this way, but it may still be worth discussing.