@astrofrog 2017-12-01
Ok and just to be clear, if we speed up ``SkyCoord.__init__`` by 20x in future, can we remove this and not break backward-compatibility? (I think yes?)