@mhvk 2017-12-01
@astrofrog, @taldcroft - I'm trying to put together a hack that just skips `__init__` for cases like these, so we have something to compare against.  But it is near the end of term here, and I have very little time. A week from now things should have quieted down a bit.