@taldcroft 2017-12-01
>  if we speed up SkyCoord.__init__ by 20x in future, can we remove this and not break backward-compatibility? 

Yes, this PR introduces no change in public API.  Obviously it adds a private attribute and method, but these can be removed in the future with no notice.