@mhvk 2017-12-07
I now tried cleaning up getting items, making sure one doesn't pass through deadly slow `__init__` - I get a factor of nearly 50 from that! See #6941. There's definitely more to be done, but even now I think it proves we don't have to add extra hacks to try to paper over not-quite-smart code (quite a bit of it my own...)