@eteq 2017-12-08
@mhvk - can you clarify : if you combine this with #6940 / #6941, do you get an *additional* speedup, or do thus essentially replace this? (my quick once-over suggests it *does*, but should be checked to determine whether this is worth keeping at least temporarily).

Assuming that is the case, I think we hold onto this one in case problems appear in #6940, as I think this is a good thing to have *if* it's not replaced by the other two.

Personally, I consider all this work unfair right now though.  What's the point of having a performance-focused release in 3.1 if we do it all now? 😉