@taldcroft 2017-12-08
It looks like this will be superceded by #6940 / #6941 if they work out.  Great work @mhvk!