@taldcroft 2017-12-22
Timings on my mac laptop show that this PR still increases getitem speed by a factor of more than 11 after merging #6940 (which itself improved speed by a little less than a factor of two).  Unless #6941 gets merged then the coordinates team should still consider this one.

**Astropy 2.0**
```
In [2]: sc = SkyCoord(np.linspace(0, 10, 100), np.linspace(0, 10, 100), unit='deg')

In [3]: timeit list(sc)
302 ms ± 9.44 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

**current master (with #6940 merged)**
```
In [8]: timeit list(sc)
10 loops, best of 3: 176 ms per loop
```

**With this PR (rebased on current master with #6940)
```
In [3]: timeit list(sc)
100 loops, best of 3: 15 ms per loop
```