@eteq 2017-12-23
Went ahead and merge #6941, so this one won't go in to 3.0.

More generally, I think we are agreed that we don't want to implement too much caching until we've taken a first crack at cleaning up the `SkyCoord` initializer and deciding *where* the caching should go.  With that in mind, @taldcroft, do you think we should leave this open for consideration in that process, or do you think this should be closed?